Add cast to eliminate warning
git-svn-id: http://svn.research.cc.gatech.edu/kaos/atl/trunk@3231 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ attr.c
@@ -209,7 +209,7 @@ attr_value *value_p;
 		if ((sizeof(long) != 4) && 
 		    (list->l.list.attributes[index].val_type == Attr_Int4)){
 		    
-		    *((int*)value_p) = (int)list->l.list.attributes[index].value;
+		    *((int*)value_p) = (int)(long)list->l.list.attributes[index].value;
 		} else {
 		    *value_p = list->l.list.attributes[index].value;
 		}
